dogs/app/core/config.py
from typing import List, Literal, Optional
from pydantic_settings import BaseSettings, SettingsConfigDict
from pydantic import SecretStr


class Settings(BaseSettings):
    TG_TOKEN: SecretStr
    OPENAI_API_KEY: SecretStr
    ADMIN_IDS: List[int] = []
    LLM_PROVIDER: Literal["openai", "anthropic"] = "openai"
    OPENAI_MODEL: str = "gpt-4o-mini"
    OPENAI_MAX_OUTPUT_TOKENS: int = 700
    ANTHROPIC_API_KEY: Optional[SecretStr] = None
    ANTHROPIC_MODEL: str = "claude-3-haiku-20240307"
    ANTHROPIC_MAX_OUTPUT_TOKENS: int = 700

    model_config = SettingsConfigDict(
        env_file=".env", env_file_encoding="utf-8", extra="ignore"
    )


settings = Settings()
