dogs/requirements.txt
aiogram>=3.0
tortoise-orm>=0.20.0
aerich>=0.7.2
pydantic-settings>=2.0
structlog
python-dotenv
aiosqlite
tomlkit
openai>=1.0.0
yt-dlp>=2023.0.0
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
tiktoken>=0.5.0
lxml>=4.9.0
anthropic>=0.29.0
